Cache Model Relationships
Reviewing trace data in newrelic revealed that models are refetched each time
a new model references a relationship. That means if there are 35 team members
and we use the center they belong to for each, the center is queried from the db
35 times.

This change addes a layer to cache the object based on type and id.