Add some caching to the CenterStatsController
When we get the global report, there are a lot of duplicate queries run
fetching the center stats data for all of the centers.

Save the global reports and search local data instead of running more queries.

Also, added memcache busting of old statsReports data when a new report is submitted