Only email this quarter's accountables
If a statistician submits their sheet in the first week and it didn't pass validation, we won't have the accurate list of accountables for the new quarter The end result is we send the report to all of last quarter's accountables.

In this case, only send the email to the center's stats email, and request they manually copy the relevant accountables.