Trim names and other things.
A number of bugs have come up in reports, transfer check, and other
places where we rely on name matching, due to the fact that often due
to cut&paste, there's additional whitespace around a name.

Use the 'trim' option to trim first/last names and other things which
are likely to have whitespace cruft.